+++ ToDoListApp/ToDoListApp/ToDoController.swift
@@ -11,10 +11,19 @@ class ToDoController: UIViewController {
 
     @IBOutlet weak var todoTableView: UITableView!
     
+    private var todoDataSource = ToDoTableViewDataSource()
+    private var todoDelegate = ToDoTableViewDelegate()
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         
+        configureDelegate()
         configureCell()
+    }
+    
+    func configureDelegate() {
+        todoTableView.delegate = todoDelegate
+        todoTableView.dataSource = todoDataSource
     }
     
     func configureCell() {
@@ -23,30 +32,7 @@ class ToDoController: UIViewController {
         todoTableView.rowHeight = UITableView.automaticDimension
         todoTableView.estimatedRowHeight = 108
         todoTableView.separatorStyle = .none
-        
     }
 
 }
 
-extension ToDoController: UITableViewDataSource, UITableViewDelegate {
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 1
-    }
-    
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return 3
-    }
-    
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return 10
-    }
-    
-    func tableView(_ tableView: UITableView, willDisplayHeaderView view: UIView, forSection section: Int) {
-        view.tintColor = UIColor.systemGray6
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = todoTableView.dequeueReusableCell(withIdentifier: "cardCell") as! CardCell
-        return cell
-    }
-}
