+++ ToDoListApp/ToDoListApp/ToDoTableViewDataSource.swift
@@ -1,0 +1,21 @@
+
+import Foundation
+import UIKit
+
+class ToDoTableViewDataSource : NSObject, UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return 1
+    }
+    
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return 3
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: "cardCell") as? CardCell
+        else {
+            return UITableViewCell()
+        }
+        return cell
+    }
+}
