+++ ToDoListApp/ToDoListApp/ToDoTableViewDelegate.swift
@@ -1,0 +1,13 @@
+
+import Foundation
+import UIKit
+
+class ToDoTableViewDelegate: NSObject, UITableViewDelegate {
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        return 10
+    }
+    
+    func tableView(_ tableView: UITableView, willDisplayHeaderView view: UIView, forSection section: Int) {
+        view.tintColor = UIColor.systemGray6
+    }
+}
